requirements.txt
fastapi
uvicorn
python-jose[cryptography]
pydantic
pytz
requests
sqlalchemy
psycopg2-binary
python-dotenv